requirements.txt
pandas~=2.2.3
pymongo~=4.10.1
requests~=2.32.3
kafka-python~=2.0.2
pyspark~=3.5.3
elasticsearch~=8.16.0
spacy~=3.8.2
nltk~=3.9.1
torch~=2.5.1
transformers~=4.47.0
